Lab11 + 12

+++ src/main/java/com/example/splabsmarandache/persistence/BooksRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BooksRepository extends JpaRepository<Book, Integer> {
+public interface BooksRepository extends JpaRepository<Book, Long> {
 
 
 }
